+++ packages/@expo/config-plugins/src/android/__tests__/Locales-test.ts
@@ -1,16 +1,16 @@
+import { XML } from 'expo/config-plugins';
 import * as fs from 'fs';
 import { vol } from 'memfs';
+import * as nodePath from 'path';
 
-import { getDirFromFS } from './utils/getDirFromFS';
-import rnFixture from '../../plugins/__tests__/fixtures/react-native-project';
+import { readAllFiles } from '../../plugins/__tests__/fixtures/react-native-project';
+import { writeXMLAsync } from '../../utils/XML';
 import * as WarningAggregator from '../../utils/warnings';
 import { getLocales, setLocalesAsync } from '../Locales';
-import { getPbxproj } from '../utils/Xcodeproj';
-
 jest.mock('fs');
 jest.mock('../../utils/warnings');
 
-describe('iOS Locales', () => {
+describe('Android Locales', () => {
   it(`returns null if no values are provided`, () => {
     expect(getLocales({})).toBeNull();
   });
@@ -24,14 +24,12 @@ describe('iOS Locales', () => {
   });
 });
 
-describe('e2e: iOS locales', () => {
+describe('e2e: Android locales', () => {
   const projectRoot = '/app';
   beforeAll(async () => {
     vol.fromJSON(
       {
-        'ios/testproject.xcodeproj/project.pbxproj':
-          rnFixture['ios/HelloWorld.xcodeproj/project.pbxproj'],
-        'ios/testproject/AppDelegate.m': '',
+        ...readAllFiles(),
         'lang/fr.json': JSON.stringify({
           ios: {
             CFBundleDisplayName: 'french-name',
@@ -56,49 +54,62 @@ describe('e2e: iOS locales', () => {
       },
       projectRoot
     );
+    const mockXML = {
+      writeXMLAsync: ({ path, xml }) => {
+        vol.mkdirSync(nodePath.dirname(path), { recursive: true });
+        vol.writeFileSync(path, jest.requireActual('../../utils/XML').format(xml));
+      },
+    };
+    const mockJSONFile = {
+      readAsync: (path) => JSON.parse(vol.readFileSync(path).toString()),
+    };
+    jest.mock('../../utils/XML', () => mockXML);
+    jest.mock('@expo/json-file', () => mockJSONFile);
   });
 
   afterAll(() => {
     vol.reset();
   });
 
-  it('writes all the image files expected', async () => {
-    let project = getPbxproj(projectRoot);
-
-    project = await setLocalesAsync(
+  it('writes all the language files expected', async () => {
+    await setLocalesAsync(
       {
         locales: {
           fr: 'lang/fr.json',
+          'en-US': 'lang/en-US.json',
+          en: 'lang/en.json',
           // doesn't exist
           xx: 'lang/xx.json',
-
           // partially support inlining the JSON so our Expo Config type doesn't conflict with the resolved manifest type.
           es: { CFBundleDisplayName: 'spanish-name' },
-          'en-US': 'lang/en-US.json',
-          // support backwards compatibility for `locales` structure without platform keys.
-          en: 'lang/en.json',
         },
       },
-      { project, projectRoot }
+      { projectRoot }
     );
-    // Sync the Xcode project with the changes.
-    fs.writeFileSync(project.filepath, project.writeSync());
 
-    const after = getDirFromFS(vol.toJSON(), projectRoot);
-    const locales = Object.keys(after).filter((value) => value.endsWith('InfoPlist.strings'));
-    expect(locales.length).toBe(4);
-    expect(after[locales[0]]).toMatchSnapshot();
-    // Test that the inlined locale is resolved.
-    expect(after[locales[1]]).toMatch(/spanish-name/);
-    expect(after[locales[2]]).toMatchInlineSnapshot(`"CFBundleDisplayName = "us-name";"`);
-    expect(after[locales[3]]).toMatchInlineSnapshot(`
-      "CFBundleDisplayName = "us-name";
-      app_name = "us-name";"
+    expect(vol.readFileSync('/app/android/app/src/main/res/values-b+es/strings.xml').toString())
+      .toMatchInlineSnapshot(`
+      "<resources>
+        <string name="CFBundleDisplayName">spanish-name</string>
+      </resources>"
+    `);
+    // backwards compatibility
+    expect(vol.readFileSync('/app/android/app/src/main/res/values-b+en/strings.xml').toString())
+      .toMatchInlineSnapshot(`
+      "<resources>
+        <string name="CFBundleDisplayName">us-name</string>
+        <string name="app_name">us-name</string>
+      </resources>"
+    `);
+    expect(vol.readFileSync('/app/android/app/src/main/res/values-b+en+US/strings.xml').toString())
+      .toMatchInlineSnapshot(`
+      "<resources>
+        <string name="app_name">us-name</string>
+      </resources>"
     `);
 
-    // Test a warning is thrown for an invalid locale JSON file.
     expect(WarningAggregator.addWarningForPlatform).toHaveBeenCalledWith(
-      'ios',
+      'android',
       'locales.xx',
       'Failed to parse JSON of locale file for language: xx',
       'https://docs.expo.dev/guides/localization/#translating-app-metadata'